+++ packages/apollo-codegen-flow/src/printer.ts
@@ -9,14 +9,14 @@ export default class Printer {
   private printQueue: Printable[] = [];
 
   public print(): string {
-    return this.printQueue.reduce((document: string, printable) => {
+    return (this.printQueue.reduce((document: string, printable) => {
       if (typeof printable === "string") {
         return document + printable;
       } else {
         const documentPart = generate(printable).code;
         return document + this.fixCommas(documentPart);
       }
-    }, "") as string;
+    }, "") as string).trim();
   }
 
   public enqueue(printable: Printable) {