+++ packages/apollo/src/Command.ts
@@ -1,5 +1,5 @@
 import Command, { flags } from "@oclif/command";
-import * as Listr from "listr";
+import Listr from "listr";
 import { ListrTask } from "listr";
 import { parse, resolve } from "path";
 
@@ -214,10 +214,19 @@ export abstract class ProjectCommand extends Command {
   }
 
   async runTasks<Result>(
-    generateTasks: (context: ProjectContext) => ListrTask[]
+    generateTasks: (context: ProjectContext) => ListrTask[],
+    options?: Listr.ListrOptions | ((ctx: ProjectContext) => Listr.ListrOptions)
   ): Promise<Result> {
-    const tasks = await generateTasks(this.ctx!);
-    return new Listr([...this.tasks, ...tasks]).run();
+    const { ctx } = this;
+    if (!ctx) {
+      throw new Error("init must be called before trying to access this.ctx");
+    }
+
+    const tasks = await generateTasks(ctx);
+    return new Listr(
+      [...this.tasks, ...tasks],
+      typeof options === "function" ? options(ctx) : options
+    ).run();
   }
   async catch(err) {
     // handle any error from the command