+++ packages/apollo/src/generate.ts
@@ -1,5 +1,5 @@
 import { fs } from "apollo-codegen-core/lib/localfs";
-import * as path from "path";
+import path from "path";
 import { GraphQLSchema, DocumentNode, print } from "graphql";
 import URI from "vscode-uri";